lol broken checkpoint
