+++ tests/jdom/scheduler.ts
@@ -26,6 +26,13 @@ export class FastForwardScheduler implements Scheduler {
 
     protected eventQueue = new Heap<EventRecord>(eventMinComparator)
 
+    public async runForDelay(delayMs: number) {
+        const promise = new Promise(resolve => {
+            this.setTimeout(() => { resolve(undefined) }, delayMs)
+        })
+        return this.runToPromise(promise)
+    }
+
     // Runs the scheduler until (at least) some condition is met, resolving the promise
     // after processing for that instant.
     // TODO: is this the right API?
